swarm/storage: split 3 impl for memstore
